refactor: 컨트롤러에서 validation, default value 처리하도록 수정
